Update app.component.ts
Hi, this the fix I did to force the googleCallback to be called in this component and set this.loggedInCreds before the call to testAccess. I don't think the duplicate calls to google (line 95 in here and line 50 in google-signin-component.ts) are required now but I am not sure why they fixed the issue even intermittently so I have left them alone.